chore(openspec): archive 16 graduated changes with receipts, spec sync, and index registration
- Archive to openspec/changes/archive/2026-08-30-<name>/: profile-source-content-pinning,
  dsh-uniflow-profile-adapter, dsh-uniflow-run-scoped-operational-state,
  universal-ai-coder-protocol-migration, uniflow-required-skill-propagation,
  perception-navigation-row-composition-repair, greenfield-agent-runtime,
  fast-semantic-container-identity-baseline, runtime-agent-decision-adaptation,
  runtime-agent-plan-hypothesis, runtime-agent-directive-capability,
  runtime-agent-reconciliation-decision, runtime-agent-strategy-execution-loop,
  uniagent-runtimeagent-strategy-contract, runtime-scenario-knowledge-boundary-cleanup,
  trace-span-read-model
- Graduation receipts: 16x docs/decisions/<name>-graduation-decision.md; registry
  docs/decisions/index.md 220 -> 236 (strategy-contract row updated to archive path)
- Delta spec sync: 16 capabilities -> openspec/specs/<cap>/spec.md, incl.
  runtime-external-contract-baseline MODIFIED merge into existing main spec
- Projections regenerated: current-gates / latest (active 10, archived 80)
- Gates: verify-before-commit ALL PASS (C1-C15, AgentWorkflow 256 passed);
  openspec validate --all --strict 90/90; git diff --check clean

+++ src/UniClaw.Runtime/Capabilities/Perception/Semantic/V2/SemanticObservationFactProjector.cs
@@ -108,8 +108,16 @@ public static class SemanticObservationFactProjector
             source.Provenance, sequence, frame, rawText, rawClassName, rawResourceName, rawContentDescription,
             primitiveState, bounds, rawProviderType, clickable, checkable, enabled, focusable, parentOccurrenceId));
 
+    /// <summary>
+    /// Normalizes a float32 element bound into double-precision semantic bounds.
+    /// The subtraction MUST happen after widening to double: float32
+    /// (1.0f − 0.002778f) rounds UP to 0.9972220063209534f, and re-widening to
+    /// double makes the fit check see left+width = 1.0000000063 > 1 for a VALID
+    /// full-width element (X2 == 1.0f, IsValid == true). Widening first makes
+    /// left+width reconstruct X2 exactly, so the invariant check is exact.
+    /// </summary>
     private static SemanticNormalizedBounds Normalize(ElementBounds bounds) =>
-        new(bounds.X1, bounds.Y1, bounds.X2 - bounds.X1, bounds.Y2 - bounds.Y1);
+        new((double)bounds.X1, (double)bounds.Y1, (double)bounds.X2 - bounds.X1, (double)bounds.Y2 - bounds.Y1);
 
     /// <summary>Creates the stable observation-local occurrence identifier used by this projector.</summary>
     public static string CreateOccurrenceId(string sourceId, string sourceOccurrenceIdentity)
